+++ Demo/Wrapper.build.ps1
@@ -177,7 +177,7 @@ Summary,
 @{TreeCyclicReference=1},
 {
 	$e = error TreeTaskNotDefined
-	assert ("$e" -like "Task 'task1': Task 'missing' is not defined.*At *z\test.build.ps1:2 *")
+	assert ("$e" -like "Task 'task1': Missing task 'missing'.*At *z\test.build.ps1:2 *")
 
 	$e = error TreeCyclicReference
 	assert ("$e" -like "Task 'task2': Cyclic reference to 'task1'.*At *z\test.build.ps1:3 *")