+++ Invoke-Build.ps1-Help.ps1
@@ -7,7 +7,7 @@
 ### Invoke-Build.ps1 command help
 @{
 	command = 'Invoke-Build.ps1'
-	synopsis = 'Invoke-Build - Build Automation in PowerShell'
+	synopsis = 'Invoke-Build - PowerShell Task Scripting'
 	description = @'
 	Install: copy Invoke-Build.ps1 and Invoke-Build.ps1-Help.xml to the path.
 
@@ -325,9 +325,6 @@
 		the Option to it. The only supported now option value is 1: protected
 		task call. It tells to ignore task errors if other active tasks also
 		call TaskName as protected.
-
-		It is fine to mix single jobs and dynamically created collections of
-		jobs in the same argument. Collection items are unrolled internally.
 '@
 		After = @'
 		Tells to add this task to job lists of the specified tasks. It is added